requirements.txt
pandas==1.5.2
psycopg2_binary==2.9.5
pytest==7.2.0
PyYAML==6.0
SQLAlchemy==1.4.45
